Module capnp_futures::serialize

Asynchronous reading and writing of messages using the standard stream framing.

Structs

OwnedSegments

Read
An in-progress read operation.

Transport

Write
An in-progress write operation.

Traits

AsOutputSegments
Something that contains segments ready to be written out.

Functions

read_message
Begins an asynchronous read of a message from reader.

write_message
Begins an asynchronous write of provided message to writer.